Add sorting ability to children for a set of args
The sorting will work for a specific set of arguments given to the
sorting file.

There is still an odd error where the column that is in question will
have a character ppended to the end. This is most likely a holdover from
the previous implementation which avoided duplicates by appeding a
character to the end. Look more into this bug.

For now this will successfully sort usiong the calls to children. What
is left to be done is implmeenting other arguments anbd failing
gracefully.