Add functionality for outputing to custom directory
The program in this state will work or 5 and 7 specified arguments. So it can output the sorted files to the current directory if an output is not specified. This is accomplished by checking file names to ensure that sorted ones are not resorted and a new directory is created.

What is left is dealing with nested directory issues and accounting for only 3 arguments. Edge case and EC should also be done.