Come closer to solving nested directories
To solve the problem look in the sorter2.c file at line 110. In here the currentPath is used to open a new directory with the updated path. This path is the next step in the hierarchy. This string is then freed.

However the directory pointer that is then thrown to the begning of the while loop on line 67 is does not have this next step in the directory.

So instead of opening ./adir/bdir/filename.csv, it will open ./adir/filename.csv

We must somehow ensure that the directory pointer from the child if statement is brought correctly to the beginnning of the while loo with the updated d_name as well.