Fix issues with nested directories
This builod is able to fork through nexted directories and output sorted csv files. In this tsate the project is turn in ready but there is a good amount of code cleanup and documentation that needs to be done.

As of now this would certainly pass so focusing on code legibility, documentation and extra credit will go a long way.