Fix segfault issue of parent
The parent process was never printing out the child_id properly and the print the ID directly to STDOUT.

The childPids array has thus been removed from the code entirely, although some comments may contain references to it that must be cleaned up. The header files and such should be updated. Look in previous commits if we want to use those child pids in some other way and store them.