Have cascading sort nearly implemented
This implemntaion is attempting to follow a cascading sort on the columns. It is proving tricky but in this state the program will sort based off of the first option, however it will not progress to sort based on the next input. More work/debugging should be done in mergesort.c

This should live exclusively on a separate branch and be handed in seperately from the main project.