Allow for cascading sort across multiple columns
Instead of iterating from 0 to 2, the looping structure is now based on the size of the columnIndeces array in sorter.c. The mergesort had to also have an additional check added to prevent an infinite loop. Essentially if all of the values for the input are the same, (i.e same actor between movies in the second level of the cascade), then it will default to sorting them as if the compare was < 0.