Output correct values and stack
The stakc that was implemented in class was thrown in here. It is not fully operational as can be seen from line 219, the stack should be filled at the point but it is said to be empty. This might have something to do with forked process memeory sharing and thread creation may already solve this.

Process counting is fully implemented and should work.