Demonstrate the stack in a forkin setting
So the stack that is implemnted here should work fine when using threads. In this implemnetation using forking, each process is essentially givem its own stack. Running this program demonstrates this if you were to run it on just a single directory of csv files. Each fork has its own stack of size = 1, so popping a single value revels that the stack is empty. Hoever in a threaded implementation the stack would not be empty as all threads can see this single stack, instead of each thread having its own stack.